Move User models to shared location
