+++ src/app/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './user.models';
