+++ src/app/users/index.ts
@@ -1,3 +1,2 @@
-export * from './models';
 export * from './users.module';
 export * from './users.service';
