+++ src/app/users/users.controller.ts
@@ -17,7 +17,7 @@ import {
 import { FastifyRequest } from 'fastify';
 import { AuthenticationGuard } from '../authentication';
 import { DmaLogger } from '../logging';
-import { CreateUserData, UpdateUserData, User } from './models';
+import { CreateUserData, UpdateUserData, User } from '../shared';
 import { UsersService } from './users.service';
 
 @UseGuards(AuthenticationGuard)
